Refactor delta-compressed data into sai2 translation unit
Move more utility functions out of the thumbnail sample
